[wip] refactor containers
