+++ universal/containers/Counter/index.js
@@ -1,23 +1,25 @@
 import { injectReducer } from 'store/reducers'
 
-export default (store) => ({
-  /*  Async getComponent is only invoked when route matches   */
-  getComponent (nextState, cb) {
-    /*  Webpack - use 'require.ensure' to create a split point
-        and embed an async module loader (jsonp) when bundling   */
-    require.ensure([], (require) => {
-      /*  Webpack - use require callback to define
-          dependencies for bundling   */
-      const Counter = require('./container').default
-      const reducer = require('../../modules/counter').default
+export default (store) => {
+  console.log('HERE')
+  return {
+    path: 'counter',
 
-      /*  Add the reducer to the store on key 'counter'  */
-      injectReducer(store, { key: 'counter', reducer })
+    getComponent (nextState, cb) {
+      console.log('getComponent Method')
+      require.ensure([], (require) => {
+        const Counter = require('./container').default
+        const module = require('../../modules/counter').default
 
-      /*  Return getComponent   */
-      cb(null, Counter)
+        console.log(module.KEY, module.reducer)
 
-    /* Webpack named bundle   */
-    }, 'counter')
+        injectReducer(store, { key: module.KEY, reducer: module.reducer })
+
+        /*  Return getComponent   */
+        cb(null, Counter)
+
+      /* Webpack named bundle   */
+      }, 'counter')
+    }
   }
-})
+}
